Refactor some reusable utilities into one place
