+++ src/flynt/string_concat/transformer.py
@@ -1,37 +1,9 @@
 import ast
 from typing import List, Tuple
 
-import astor
-
-from flynt.exceptions import FlyntException
-from flynt.format import QuoteTypes, set_quote_type
 from flynt.string_concat.candidates import is_string_concat
 from flynt.string_concat.string_in_string import check_sns_depth
-from flynt.linting.fstr_lint import FstrInliner
-
-
-def ast_formatted_value(
-    val, fmt_str: str = None, conversion=None
-) -> ast.FormattedValue:
-    if isinstance(val, ast.FormattedValue):
-        return val
-
-    if astor.to_source(val)[0] == "{":
-        raise FlyntException(
-            "values starting with '{' are better left not transformed."
-        )
-
-    if fmt_str:
-        format_spec = ast.JoinedStr([ast_string_node(fmt_str.replace(":", ""))])
-    else:
-        format_spec = None
-
-    conversion = -1 if conversion is None else ord(conversion.replace("!", ""))
-    return ast.FormattedValue(value=val, conversion=conversion, format_spec=format_spec)
-
-
-def ast_string_node(string: str) -> ast.Str:
-    return ast.Str(s=string)
+from flynt.utils import ast_formatted_value, ast_string_node, fixup_transformed
 
 
 def unpack_binop(node: ast.BinOp) -> List[ast.AST]:
@@ -94,15 +66,6 @@ def transform_concat(code: str, *args, **kwargs) -> Tuple[str, bool]:
 
     ft = ConcatTransformer()
     ft.visit(tree)
-    il = FstrInliner()
-    il.visit(tree)
-
-    new_code = astor.to_source(tree)
-    if new_code[-1] == "\n":
-        new_code = new_code[:-1]
-
-    new_code = new_code.replace("\n", "\\n")
-    if new_code[:4] == 'f"""':
-        new_code = set_quote_type(new_code, QuoteTypes.double)
+    new_code = fixup_transformed(tree)
 
     return new_code, ft.counter > 0
